pyproject.toml
[project]
name = "patient-discharge-education"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
